remove empty file
